+++ src/index.ts
@@ -3,6 +3,7 @@ export { Checkbox } from './components/checkbox';
 export { CheckboxGroup } from './components/checkbox-group';
 export { Loader } from './components/loader';
 export { Button } from './components/button';
+export { ButtonGroup } from './components/button-group';
 export { IconBarChart } from './components/icons/bar-chart';
 export { IconBook } from './components/icons/book';
 export { IconBookmark } from './components/icons/bookmark';